requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
aiohttp==3.9.1
python-jose==3.3.0
cryptography==41.0.7 